backend/pyproject.toml
[project]
name = "gluemigrator"
version = "0.1.0"
description = "Migration app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "cryptography>=42.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
